Rename ProjectGridTemplate
... to ProjectGrid
